Move useInternal hook
